+++ R/FilteredDataset-utils.R
@@ -82,25 +82,25 @@
 #' }
 #' @keywords internal
 #' @export
-init_filtered_dataset <- function(dataset, # nolint
+init_filtered_dataset <- function(dataset,
                                   dataname,
                                   keys = character(0),
                                   parent_name = character(0),
                                   parent = reactive(dataset),
                                   join_keys = character(0),
-                                  label = attr(dataset, "label")) {
+                                  label = attr(dataset, "label", exact = TRUE)) {
   UseMethod("init_filtered_dataset")
 }
 
 #' @keywords internal
 #' @export
-init_filtered_dataset.data.frame <- function(dataset, # nolint
+init_filtered_dataset.data.frame <- function(dataset,
                                              dataname,
                                              keys = character(0),
                                              parent_name = character(0),
                                              parent = NULL,
                                              join_keys = character(0),
-                                             label = attr(dataset, "label")) {
+                                             label = attr(dataset, "label", exact = TRUE)) {
   DataframeFilteredDataset$new(
     dataset = dataset,
     dataname = dataname,
@@ -114,13 +114,13 @@ init_filtered_dataset.data.frame <- function(dataset, # nolint
 
 #' @keywords internal
 #' @export
-init_filtered_dataset.MultiAssayExperiment <- function(dataset, # nolint
+init_filtered_dataset.MultiAssayExperiment <- function(dataset,
                                                        dataname,
                                                        keys = character(0),
                                                        parent_name, # ignored
                                                        parent, # ignored
                                                        join_keys, # ignored
-                                                       label = attr(dataset, "label")) {
+                                                       label = attr(dataset, "label", exact = TRUE)) {
   if (!requireNamespace("MultiAssayExperiment", quietly = TRUE)) {
     stop("Cannot load MultiAssayExperiment - please install the package or restart your session.")
   }
@@ -134,13 +134,13 @@ init_filtered_dataset.MultiAssayExperiment <- function(dataset, # nolint
 
 #' @keywords internal
 #' @export
-init_filtered_dataset.default <- function(dataset, # nolint
+init_filtered_dataset.default <- function(dataset,
                                           dataname,
                                           keys, # ignored
                                           parent_name, # ignored
                                           parent, # ignored
                                           join_keys, # ignored
-                                          label = attr(dataset, "label")) {
+                                          label = attr(dataset, "label", exact = TRUE)) {
   DefaultFilteredDataset$new(
     dataset = dataset,
     dataname = dataname,