+++ NEWS.md
@@ -12,6 +12,10 @@
 * `set_filter_state` no longer accepts a nested list. Use `teal_slices()` and `teal_slice()` instead.
 * Renamed `FilteredDataset` subclass that handles `data.frame`s from `DefaultFilteredDataset` to `DataframeFilteredDataset`. Added new class `DefaultFilteredDataset` that will store any type of object. Filtering will is not supported.
 
+### Bug fixes
+
+* Performs an exact match when determining the default label of a dataset from attributes.
+
 ### Miscellaneous
 
 * Specified minimal version of package dependencies.