+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.slice 0.5.0.9002
+# teal.slice 0.5.0.9003
 
 # teal.slice 0.5.0
 
@@ -11,6 +11,10 @@
 * `TealData` object is no longer supported by `teal.slice`.
 * `set_filter_state` no longer accepts a nested list. Use `teal_slices()` and `teal_slice()` instead.
 * Renamed `FilteredDataset` subclass that handles `data.frame`s from `DefaultFilteredDataset` to `DataframeFilteredDataset`. Added new class `DefaultFilteredDataset` that will store any type of object. Filtering will is not supported.
+
+### Bug fixes
+
+* Performs an exact match when determining the default label of a dataset from attributes.
 
 ### Miscellaneous